+++ src/main/java/day13/Memory.java
@@ -4,29 +4,29 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.regex.Pattern;
 
-import static java.lang.Long.parseLong;
+import static java.lang.Integer.parseInt;
 import static java.util.stream.IntStream.range;
 
 class Memory {
 
-    private final Map<Long, Long> storage = new HashMap<>();
+    private final Map<Integer, Integer> storage = new HashMap<>();
 
     public Memory(String program) {
         var values = Pattern.compile(",").split(program);
-        range(0, values.length).forEach(i -> storage.put((long) i, parseLong(values[i])));
+        range(0, values.length).forEach(i -> storage.put(i, parseInt(values[i])));
     }
 
-    public long read(long address) {
+    public int read(int address) {
         validateAddress(address);
-        return storage.getOrDefault(address, 0L);
+        return storage.getOrDefault(address, 0);
     }
 
-    public void write(long address, long value) {
+    public void write(int address, int value) {
         validateAddress(address);
         storage.put(address, value);
     }
 
-    private void validateAddress(long address) {
+    private void validateAddress(int address) {
         if (address < 0) {
             throw new IllegalArgumentException("invalid address: " + address);
         }
